+++ data/sample_data_inflow/mock_motif_upload.py
@@ -1,0 +1,54 @@
+import os
+import csv
+import random
+from database.models import SectionClass
+from data_inflow.motifs_upload import MotifUploader
+
+
+def create_mock_motif_batch():
+    BATCH_SIZE = 10000
+    TEMP_CSV_PATH = "data/sample_data/temp_stress_test_motif.csv"
+    TARGET_TRACK_ID = 1
+
+    # Creates the temporary output path for the CSV batch file
+    os.makedirs(os.path.dirname(TEMP_CSV_PATH), exist_ok=True)
+
+    print(f"Generating motif batch CSV for {BATCH_SIZE} motifs...")
+
+    # Initialize list of eligible motif classes based on Enum labels
+    section_classes = list(SectionClass)
+
+    # Opens the CSV file in 'write' mode, adding a dummy sequence of notes to rows pertaining to each mock motif
+    with open(TEMP_CSV_PATH, mode='w', newline='') as file:
+        writer = csv.writer(file)
+        writer.writerow(['motif_name', 'motif_class', 'pitch_value', 'beat_position', 'duration'])
+
+        for i in range(BATCH_SIZE):
+            m_name = f"STRESS_TEST_{i}"
+            m_class = random.choice(section_classes).value # Song structure blocks are selected at random
+
+            writer.writerow([m_name, m_class, random.randint(60, 72), 0.0, 1.0])
+            writer.writerow([m_name, m_class, random.randint(60, 72), 1.0, 1.0])
+            writer.writerow([m_name, m_class, random.randint(60, 72), 2.0, 1.0])
+            writer.writerow([m_name, m_class, random.randint(60, 72), 3.0, 1.0])
+    
+    print("CSV generation complete. Passing to CSVUploader pipeline...")
+
+    # Initializes path to Motif CSV Uploader, designating the required information needed to upload the mock motif batch
+    uploader = MotifUploader()
+    result = uploader.upload_batch(
+        csv_file_path = TEMP_CSV_PATH,
+        track_id = TARGET_TRACK_ID
+    )
+
+    if result['status'] == 'success':
+        print(result['message'])
+    else:
+        print(f"Pipeline failure: {result['message']}")
+    
+    # Deletes the temporary file so leftover data doesn't corrupt the next test.
+    if os.path.exists(TEMP_CSV_PATH):
+        os.remove(TEMP_CSV_PATH)
+
+if __name__ == "__main__":
+    create_mock_motif_batch()