+++ database/models.py
@@ -141,6 +141,7 @@ class Motif(Base):
     motif_name = Column(String(50), nullable=False)             # Name label of current motif
     sequence_data: str = Column(Text, nullable=False)           # type: ignore (Taxonomy Token to classify motifs using their IDs and Enum labels)
     motif_class = Column(Enum(SectionClass), nullable=False)    # Motif classification used to match corresponding blueprint container label
+    motif_tag = Column(String(50))                              # Optional sub-classification for heuristic transition grouping
     phrase_latency = Column(Float, nullable=False, default=0.0, server_default="0.0")       # Applies phrase or sequence-level human timing to the motif
     motif_pivot_offset = Column(Float, nullable=False, default=0.0, server_default="0.0")   # "Boundary Line" of a motif, shorter motifs loop until this number is reached, longer motifs are cut off at this point
     created_at = Column(TIMESTAMP, server_default=func.now())   # Log for creation (Upload) timestamp of motif (Allows the DB to handle the timestamp automatically)